Refactor: Adapted for the upgraded knowledge graph of @antv/g6.
