+++ web/src/utils/list-filter-util.ts
@@ -10,14 +10,16 @@ export function groupListByType<T extends Record<string, any>>(
   labelField: string,
 ) {
   const fileTypeList: FilterType[] = [];
-  list.forEach((x) => {
-    const item = fileTypeList.find((y) => y.id === x[idField]);
-    if (!item) {
-      fileTypeList.push({ id: x[idField], label: x[labelField], count: 1 });
-    } else {
-      item.count += 1;
-    }
-  });
+  if (Array.isArray(list)) {
+    list.forEach((x) => {
+      const item = fileTypeList.find((y) => y.id === x[idField]);
+      if (!item) {
+        fileTypeList.push({ id: x[idField], label: x[labelField], count: 1 });
+      } else {
+        item.count += 1;
+      }
+    });
+  }
 
   return fileTypeList;
 }
